Move RewriteEqualityTransformer to own class in SmtLibUtils
